fixed findings from eslint, new env file for helm.dev, renamed page-file for sonarcloud
