+++ .env.dev
@@ -1,4 +1,4 @@
-UMGEBUNG="dev"
+UMGEBUNG="test.dev"
 PW=""
 USER=""
 URL_PORTAL="https://test.dev.spsh.dbildungsplattform.de/"
