+++ .env.helm
@@ -1,0 +1,4 @@
+UMGEBUNG="helm.dev"
+PW=""
+USER=""
+URL_PORTAL="https://helm.dev.spsh.dbildungsplattform.de/"
