--- .env.local
@@ -1,4 +1,0 @@
-UMGEBUNG="local"
-PW=""
-USER=""
-URL_PORTAL="https://test.dev.spsh.dbildungsplattform.de/"
