+++ pages/menue.page..ts
@@ -1,4 +1,4 @@
-import { test, expect, type Locator, type Page } from '@playwright/test';
+import { type Locator, type Page } from '@playwright/test';
 
 export class MenuePage{
     readonly page: Page;
