+++ pages/start.page.ts
@@ -1,4 +1,4 @@
-import { test, expect, type Locator, type Page } from '@playwright/test';
+import { type Page } from '@playwright/test';
 
 export class StartseitePage{
     readonly page: Page;
