+++ playwright.config.ts
@@ -1,6 +1,8 @@
 import { defineConfig, devices } from '@playwright/test';
-// require("dotenv").config({ path: "./.env.local" });
-require("dotenv").config({ path: "./.env.dev" });
+import dotenv from "dotenv";
+dotenv.config({
+  path: "./.env.dev",
+});
 
 export default defineConfig({
   testDir: './tests',
