+++ tests/logoff.spec.ts
@@ -1,7 +1,6 @@
 import { test, expect } from '@playwright/test';
-import { LoginPage } from '../pages/login';
-import { LandingPage } from '../pages/landing';
-import { MenuePage } from '../pages/menue';
+import { LoginPage } from '../pages/login.page.';
+import { MenuePage } from '../pages/menue.page.';
 
 const PW = process.env.PW;
 const USER = process.env.USER;
@@ -10,7 +9,6 @@ const URL_PORTAL = process.env.URL_PORTAL;
 test.describe(`Testfälle für die Authentifizierung: Umgebung: ${process.env.UMGEBUNG}: URL: ${process.env.URL_PORTAL}:`, () => {
   test('Erfolgreicher Standard Logoff', async ({ page }) => {
     const Login = new LoginPage(page);
-    const Landing = new LandingPage(page);
     const Menue = new MenuePage(page);
 
     await test.step(`Annmelden mit Benutzer ${USER}`, async () => {
@@ -24,5 +22,7 @@ test.describe(`Testfälle für die Authentifizierung: Umgebung: ${process.env.UM
       await Menue.button_Abmelden.click(); 
       await expect(Menue.button_Anmelden).toBeEnabled();
     })
+
+    // await page.pause();
   })  
 })
